[fix]: extracting components
